pyproject.toml
[project]
name = "eclipse-horizon-checker"
version = "0.2.0"
description = "Check terrain and tree visibility for the 2026 total solar eclipse"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "numpy==2.*",
  "matplotlib==3.*",
  "requests==2.*"
]

[project.optional-dependencies]
canopy = [
    "rasterio>=1.3",
    "Pillow>=10.0",
]

[dependency-groups]
dev = [
  "pytest==8.*",
  "mypy==1.*",
  "ruff==0.*",
  "types-requests==2.*"
]